requirements.txt
PySide6>=6.6
PySide6-Addons
PySide6-Essentials
mysql-connector-python>=8.2
pyserial>=3.5
qrcode[pil]>=7.4
Pillow>=10.0
python-dateutil